Improve code and add isActive flag to allow logic exclusion

+++ BlackBeans/Model/API/Structs/APICategory.swift
@@ -9,6 +9,6 @@
 import Foundation
 
 struct APICategory: APICodable {
-  let id: Int
+  let id: Int64
   let name: String
 }
